feat: GitHub Enterprise Server 2.18
BREAKING CHANGE: :  parameter must now be a '1' instead of 1. It's due to the new way files are generated, working around it is not worth the effort

BREAKING CHANGE:  has been removed for GHE 2.17 and below. This has been resolved with GHE 2.18. The problem is that the endpoint accepts different inputs which before resulted in two different methods. With the new OpenAPI specification of octokit/routes, this can no longer be ... specified (at least not without a lot of extra effort). So sorry for that. You can create a pull request comment using  instead: